Enforce database constraints

+++ app/src/main/java/com/example/recipeapp/ShoppingListIngredient.java
@@ -1,14 +1,14 @@
 package com.example.recipeapp;
 
-// same as ingredient, except there are no SUPERMARKET column
+// same as ingredient, except there are no SUPERMARKET column and amount is INT
 // It is separately stored in the hashmap and db.
 
 public class ShoppingListIngredient {
     private String ingredientName;
-    private float amount;
+    private int amount;
     private float cost;
 
-    public ShoppingListIngredient(String ingredientName, float amount, float cost){
+    public ShoppingListIngredient(String ingredientName, int amount, float cost){
         this.ingredientName = ingredientName;
         this.amount = amount;
         this.cost = cost;
@@ -18,7 +18,7 @@ public class ShoppingListIngredient {
         return ingredientName;
     }
 
-    public float getAmount() {
+    public int getAmount() {
         return amount;
     }
